+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v2.14] - 2025-12-03
 
 ## Added 
+- `MAX_DENOMINATOR` constants 
+- `NewBaseCrncyData` instruction data
+- `CommunityAccount` to `perp_upgrade`
+- Fields in `InstrAccountHeader`: `init_seat_price` and `ema_px`
 - `MIN_AMOUNT_FOR_DIVIDENDS_ALLOCATION` constant
 - `pepr_last_trade_asset/crncy_tokens` and `last_trade_asset/crncy_tokens` statistic for last trade record